fix(api): migrate action deprecated gitclone
* don't migrate if we don't find any key
* move the code in the migrate package
* run it as a goroutine
